parser re-arrangement
