+++ src/toolchain/diagnostics/diagnostic_kind.rs
@@ -1,0 +1,4 @@
+/// The enumerated type of all diagnostics Hadron emits.
+pub enum DiagnosticKind {
+    ParseUnknownToken,
+}
