+++ src/toolchain/mod.rs
@@ -1,3 +1,3 @@
-//pub mod diagnostics;
+pub mod diagnostics;
 pub mod lexer;
 //pub mod parser;
