+++ src/toolchain/parser/node_kind.rs
@@ -1,0 +1,43 @@
+pub enum NodeKind {
+    ArgList,
+    ArrayRead,
+    ArrayWrite,
+    Assign,
+    BinopCall,
+    Block,
+    Call,
+    Class,
+    ClassExtension,
+    Collection,
+    CopySeries,
+    Empty,
+    EnvironmentAt,
+    EnvironmentPut,
+    Event,
+    ExprSeq,
+    KeyValue,
+    Method,
+    MultiAssign,
+    MultiAssignVars,
+    Name,
+    New,
+    NumericSeries,
+    PerformList,
+    PutSeries,
+    Return,
+    Series,
+    SeriesIter,
+    Setter,
+    Slot,
+    String,
+    Symbol,
+    Value,
+    VarDef,
+    VarList,
+    ListComprehension,
+    GeneratorQualifier,
+    GuardQualifier,
+    BindingQualifier,
+    SideEffectQualifier,
+    TerminationQualifier
+}
